Reorganising folder structure.
