+++ gui/icons/arrow-right.png
