+++ gui/icons/home.png
