+++ gui/popupWidgets.py
@@ -12,7 +12,7 @@ import datetime
 class irodsCreateCollection(QDialog):
     def __init__(self, parent, ic):
         super(irodsCreateCollection, self).__init__()
-        loadUi("ui-files/createCollection.ui", self)
+        loadUi("gui/ui-files/createCollection.ui", self)
         self.setWindowTitle("Create iRODS collection")
         self.ic = ic
         self.parent = parent
@@ -35,7 +35,7 @@ class irodsCreateCollection(QDialog):
 class createDirectory(QDialog):
     def __init__(self, parent):
         super(createDirectory, self).__init__()
-        loadUi("ui-files/createCollection.ui", self)
+        loadUi("gui/ui-files/createCollection.ui", self)
         self.setWindowTitle("Create directory")
         self.parent = parent
         self.label.setText(self.parent+os.sep)
@@ -57,7 +57,7 @@ class createDirectory(QDialog):
 class irodsIndexPopup(QDialog):
     def __init__(self, irodsTarIndexFileList, tarFilePath, statusLabel):
         super(irodsIndexPopup, self).__init__()
-        loadUi("ui-files/irodsIndexPopup.ui", self)
+        loadUi("gui/ui-files/irodsIndexPopup.ui", self)
         self.setWindowTitle("iRODS Tar/Zip index.")
         self.indexLabel.setText("Index of " + tarFilePath + ":")
         self.closeButton.clicked.connect(self.closeWindow)
