+++ gui/ui-files/createCollection.ui
