+++ gui/ui-files/searchDialog.ui
