--- iBridgesUploadClient/elabConfig
@@ -1,11 +1,0 @@
-[iRODS]
-irodsenv = /<path to>/irods_environment.json
-irodscoll = 
-irodsresc = 
-webdav = http://<webdav server fqdn>
-
-[ELN]
-token = wur.elabjournal.com;<some hash>
-group =
-experiment =
-title =
