+++ rules/tarCollection.r
