+++ rules/tarReadIndex.r
